Create 3D sensitivity map
This was part of my project in the Amgen Scholars Program in University of Tokyo in summer 2017.

I used Igor software to interface with a system of actuators(3 linear, 1 rotational). I was given the previous Amgen Scholar's code to build on. However, her code was essentially just a 2D sensitivity map in the shape of a square. That being said, as I was not familiar with Igor software, it was a good guide on the various functions relevant to this project.

This aim of this project was to produce a 3D magnetic sensitivity map for a handheld magnetic probe developed by the Sekino laboratory of the University of Tokyo. The probe is to be used during partial mastectomy surgery in order to locate tumors/lesions in the extracted tissue. Using this map, I was to then produce algorithms to determine the location of magnetic material in a given tissue sample. 
Along with the program, I was also tasked with creating a mechanical probe platform that would hold the tissue in place while having 3 types of motion (rotational, horizontal, vertical). As the platform was circular in shape, I designed the data collection to be in  a semi-circular polar plot. 

As the system was not calibrated perfectly, a lot of work was done to get the numbers right regarding step increments.